refactor(ext/node): consolidate node:fs modules (part 5)
Inline `_fs_mkdir.ts`, `_fs_mkdtemp.ts`, `_fs_open.ts`, and `_fs_opendir.ts`
into `fs.ts` and `internal/fs/promises.ts`, continuing the consolidation series.
