+++ ext/node/lib.rs
@@ -387,10 +387,6 @@ deno_core::extension!(deno_node,
     "_fs/_fs_glob.ts",
     "_fs/_fs_lstat.ts",
     "_fs/_fs_lutimes.ts",
-    "_fs/_fs_mkdir.ts",
-    "_fs/_fs_mkdtemp.ts",
-    "_fs/_fs_open.ts",
-    "_fs/_fs_opendir.ts",
     "_fs/_fs_read.ts",
     "_fs/_fs_readdir.ts",
     "_fs/_fs_readFile.ts",
