+++ ext/node/polyfills/fs.ts
@@ -19,10 +19,6 @@ import { exists, existsSync } from "ext:deno_node/_fs/_fs_exists.ts";
 import { fstat, fstatSync } from "ext:deno_node/_fs/_fs_fstat.ts";
 import { lstat, lstatSync } from "ext:deno_node/_fs/_fs_lstat.ts";
 import { lutimes, lutimesSync } from "ext:deno_node/_fs/_fs_lutimes.ts";
-import { mkdir, mkdirSync } from "ext:deno_node/_fs/_fs_mkdir.ts";
-import { mkdtemp, mkdtempSync } from "ext:deno_node/_fs/_fs_mkdtemp.ts";
-import { open, openSync } from "ext:deno_node/_fs/_fs_open.ts";
-import { opendir, opendirSync } from "ext:deno_node/_fs/_fs_opendir.ts";
 import { read, readSync } from "ext:deno_node/_fs/_fs_read.ts";
 import { readdir, readdirSync } from "ext:deno_node/_fs/_fs_readdir.ts";
 import { readFile, readFileSync } from "ext:deno_node/_fs/_fs_readFile.ts";
@@ -59,21 +55,24 @@ import {
   getValidMode,
   kMaxUserId,
   type RmOptions,
+  stringToFlags,
   toUnixTimestamp as _toUnixTimestamp,
   validateRmdirOptions,
   validateRmOptions,
   validateRmOptionsSync,
+  warnOnNonPortableTemplate,
 } from "ext:deno_node/internal/fs/utils.mjs";
 import { glob, globSync } from "ext:deno_node/_fs/_fs_glob.ts";
 import {
   parseFileMode,
+  validateBoolean,
   validateFunction,
   validateInt32,
   validateInteger,
   validateObject,
   validateString,
 } from "ext:deno_node/internal/validators.mjs";
-import type { Buffer } from "node:buffer";
+import { Buffer } from "node:buffer";
 import {
   op_fs_fchmod_async,
   op_fs_fchmod_sync,
@@ -84,6 +83,10 @@ import {
   op_node_lchmod_sync,
   op_node_lchown,
   op_node_lchown_sync,
+  op_node_mkdtemp,
+  op_node_mkdtemp_sync,
+  op_node_open,
+  op_node_open_sync,
   op_node_rmdir,
   op_node_rmdir_sync,
 } from "ext:core/ops";
@@ -91,9 +94,13 @@ import { FsFile } from "ext:deno_fs/30_fs.js";
 import {
   ERR_FS_RMDIR_ENOTDIR,
   ERR_INVALID_ARG_TYPE,
+  uvException,
 } from "ext:deno_node/internal/errors.ts";
 import { toUnixTimestamp } from "ext:deno_node/internal/fs/utils.mjs";
-import { isMacOS } from "ext:deno_node/_util/os.ts";
+import { isMacOS, isWindows } from "ext:deno_node/_util/os.ts";
+import { normalizeEncoding } from "ext:deno_node/internal/util.mjs";
+import { resolve, toNamespacedPath } from "node:path";
+import type { Encoding } from "node:crypto";
 import { core, primordials } from "ext:core/mod.js";
 
 const {
@@ -104,6 +111,7 @@ const {
   NumberIsNaN,
   ObjectPrototypeIsPrototypeOf,
   PromisePrototypeThen,
+  StringPrototypeToString,
   SymbolFor,
 } = primordials;
 
@@ -864,6 +872,452 @@ function rmdirSync(path: string | Buffer | URL, options?: rmdirOptions) {
     op_node_rmdir_sync(path);
   } catch (err) {
     throw (denoErrorToNodeError(err as Error, { syscall: "rmdir", path }));
+  }
+}
+
+// -- mkdir --
+
+type MkdirCallback =
+  | ((err: Error | null, path?: string) => void)
+  | CallbackWithError;
+
+/**
+ * On Windows, recursive mkdir through a file returns EEXIST instead of
+ * ENOTDIR. Check if any component of the path is a file and fix the error.
+ */
+function fixMkdirError(
+  err: Error,
+  path: string,
+): Error {
+  const nodeErr = denoErrorToNodeError(err, { syscall: "mkdir", path });
+  if (!isWindows) return nodeErr;
+  if ((nodeErr as NodeJS.ErrnoException).code !== "EEXIST") return nodeErr;
+  let cursor = resolve(path, "..");
+  while (true) {
+    try {
+      const stat = Deno.statSync(cursor);
+      if (!stat.isDirectory) {
+        return uvException({
+          errno: codeMap.get("ENOTDIR")!,
+          syscall: "mkdir",
+          path,
+        });
+      }
+      break;
+    } catch {
+      const parent = resolve(cursor, "..");
+      if (parent === cursor) break;
+      cursor = parent;
+    }
+  }
+  return nodeErr;
+}
+
+/** Find the first component of `path` that does not exist. */
+function findFirstNonExistent(path: string): string | undefined {
+  let cursor = resolve(path);
+  while (true) {
+    try {
+      Deno.statSync(cursor);
+      return undefined;
+    } catch {
+      const parent = resolve(cursor, "..");
+      if (parent === cursor) {
+        return toNamespacedPath(cursor);
+      }
+      try {
+        Deno.statSync(parent);
+        return toNamespacedPath(cursor);
+      } catch {
+        cursor = parent;
+      }
+    }
+  }
+}
+
+type MkdirOptions =
+  | { recursive?: boolean; mode?: number | undefined }
+  | number
+  | boolean;
+
+function mkdir(
+  path: string | URL,
+  options?: MkdirOptions | MkdirCallback,
+  callback?: MkdirCallback,
+) {
+  path = getValidatedPath(path) as string;
+
+  let mode = 0o777;
+  let recursive = false;
+
+  if (typeof options == "function") {
+    callback = options;
+  } else if (typeof options === "number") {
+    mode = parseFileMode(options, "mode");
+  } else if (typeof options === "boolean") {
+    recursive = options;
+  } else if (options) {
+    if (options.recursive !== undefined) recursive = options.recursive;
+    if (options.mode !== undefined) {
+      mode = parseFileMode(options.mode, "options.mode");
+    }
+  }
+  validateBoolean(recursive, "options.recursive");
+
+  let firstNonExistent: string | undefined;
+  try {
+    firstNonExistent = recursive ? findFirstNonExistent(path) : undefined;
+  } catch (err) {
+    if (typeof callback === "function") {
+      callback(
+        denoErrorToNodeError(err as Error, { syscall: "mkdir", path }),
+      );
+    }
+    return;
+  }
+
+  PromisePrototypeThen(
+    Deno.mkdir(path, { recursive, mode }),
+    () => {
+      if (typeof callback === "function") {
+        callback(null, firstNonExistent);
+      }
+    },
+    (err: Error) => {
+      if (typeof callback === "function") {
+        callback(
+          recursive
+            ? fixMkdirError(err as Error, path as string)
+            : denoErrorToNodeError(err as Error, { syscall: "mkdir", path }),
+        );
+      }
+    },
+  );
+}
+
+function mkdirSync(
+  path: string | URL,
+  options?: MkdirOptions,
+): string | undefined {
+  path = getValidatedPath(path) as string;
+
+  let mode = 0o777;
+  let recursive = false;
+
+  if (typeof options === "number") {
+    mode = parseFileMode(options, "mode");
+  } else if (typeof options === "boolean") {
+    recursive = options;
+  } else if (options) {
+    if (options.recursive !== undefined) recursive = options.recursive;
+    if (options.mode !== undefined) {
+      mode = parseFileMode(options.mode, "options.mode");
+    }
+  }
+  validateBoolean(recursive, "options.recursive");
+
+  let firstNonExistent: string | undefined;
+  try {
+    firstNonExistent = recursive ? findFirstNonExistent(path) : undefined;
+    Deno.mkdirSync(path, { recursive, mode });
+  } catch (err) {
+    throw recursive
+      ? fixMkdirError(err as Error, path)
+      : denoErrorToNodeError(err as Error, { syscall: "mkdir", path });
+  }
+
+  return firstNonExistent;
+}
+
+// -- mkdtemp --
+
+type MkdtempCallback = (
+  err: Error | null,
+  directory?: string,
+) => void;
+type MkdtempBufferCallback = (
+  err: Error | null,
+  directory?: Buffer<ArrayBufferLike>,
+) => void;
+
+function mkdtemp(
+  prefix: string | Buffer | Uint8Array | URL,
+  callback: MkdtempCallback,
+): void;
+function mkdtemp(
+  prefix: string | Buffer | Uint8Array | URL,
+  options: { encoding: "buffer" } | "buffer",
+  callback: MkdtempBufferCallback,
+): void;
+function mkdtemp(
+  prefix: string | Buffer | Uint8Array | URL,
+  options: { encoding: string } | string,
+  callback: MkdtempCallback,
+): void;
+function mkdtemp(
+  prefix: string | Buffer | Uint8Array | URL,
+  options: { encoding: string } | string | MkdtempCallback | undefined,
+  callback?: MkdtempCallback | MkdtempBufferCallback,
+) {
+  if (typeof options === "function") {
+    callback = options;
+    options = undefined;
+  }
+  callback = makeCallback(callback);
+  const encoding = parseMkdtempEncoding(options);
+  prefix = getValidatedPathToString(prefix, "prefix");
+
+  warnOnNonPortableTemplate(prefix);
+
+  PromisePrototypeThen(
+    op_node_mkdtemp(prefix),
+    (path: string) => callback(null, decodeMkdtemp(path, encoding)),
+    (err: Error) =>
+      callback(denoErrorToNodeError(err, {
+        syscall: "mkdtemp",
+        path: `${prefix}XXXXXX`,
+      })),
+  );
+}
+
+function mkdtempSync(
+  prefix: string | Buffer | Uint8Array | URL,
+  options?: { encoding: "buffer" } | "buffer",
+): Buffer<ArrayBufferLike>;
+function mkdtempSync(
+  prefix: string | Buffer | Uint8Array | URL,
+  options?: { encoding: string } | string,
+): string;
+function mkdtempSync(
+  prefix: string | Buffer | Uint8Array | URL,
+  options?: { encoding: string } | string,
+): string | Buffer<ArrayBufferLike> {
+  const encoding = parseMkdtempEncoding(options);
+  prefix = getValidatedPathToString(prefix, "prefix");
+
+  warnOnNonPortableTemplate(prefix);
+
+  try {
+    const path = op_node_mkdtemp_sync(prefix) as string;
+    return decodeMkdtemp(path, encoding);
+  } catch (err) {
+    throw denoErrorToNodeError(err as Error, {
+      syscall: "mkdtemp",
+      path: `${prefix}XXXXXX`,
+    });
+  }
+}
+
+function decodeMkdtemp(str: string, encoding: Encoding): string;
+function decodeMkdtemp(
+  str: string,
+  encoding: "buffer",
+): Buffer<ArrayBufferLike>;
+function decodeMkdtemp(
+  str: string,
+  encoding: Encoding | "buffer",
+): string | Buffer<ArrayBufferLike> {
+  if (encoding === "utf8") return str;
+  const buffer = Buffer.from(str);
+  if (encoding === "buffer") return buffer;
+  // deno-lint-ignore prefer-primordials
+  return buffer.toString(encoding);
+}
+
+function parseMkdtempEncoding(
+  options: string | { encoding?: string } | undefined,
+): Encoding | "buffer" {
+  let encoding: string | undefined;
+
+  if (typeof options === "undefined" || options === null) {
+    encoding = "utf8";
+  } else if (typeof options === "string") {
+    encoding = options;
+  } else if (typeof options === "object") {
+    encoding = options.encoding ?? "utf8";
+  } else {
+    throw new ERR_INVALID_ARG_TYPE("options", ["string", "Object"], options);
+  }
+
+  if (encoding === "buffer") {
+    return encoding;
+  }
+
+  const parsedEncoding = normalizeEncoding(encoding);
+  if (!parsedEncoding) {
+    throw new ERR_INVALID_ARG_TYPE("encoding", encoding, "is invalid encoding");
+  }
+
+  return parsedEncoding;
+}
+
+// -- open --
+
+type OpenFlags =
+  | "a"
+  | "ax"
+  | "a+"
+  | "ax+"
+  | "as"
+  | "as+"
+  | "r"
+  | "r+"
+  | "rs"
+  | "rs+"
+  | "w"
+  | "wx"
+  | "w+"
+  | "wx+"
+  | number
+  | string;
+
+type OpenCallback = (err: Error | null, fd?: number) => void;
+
+function open(path: string | Buffer | URL, callback: OpenCallback): void;
+function open(
+  path: string | Buffer | URL,
+  flags: OpenFlags,
+  callback: OpenCallback,
+): void;
+function open(
+  path: string | Buffer | URL,
+  flags: OpenFlags,
+  mode: number,
+  callback: OpenCallback,
+): void;
+function open(
+  path: string | Buffer | URL,
+  flags: OpenCallback | OpenFlags,
+  mode?: OpenCallback | number,
+  callback?: OpenCallback,
+) {
+  path = getValidatedPathToString(path);
+  if (arguments.length < 3) {
+    // deno-lint-ignore no-explicit-any
+    callback = flags as any;
+    flags = "r";
+    mode = 0o666;
+  } else if (typeof mode === "function") {
+    callback = mode;
+    mode = 0o666;
+  } else {
+    mode = parseFileMode(mode, "mode", 0o666);
+  }
+  flags = stringToFlags(flags);
+  callback = makeCallback(callback);
+
+  PromisePrototypeThen(
+    op_node_open(path, flags, mode),
+    (rid: number) => callback(null, rid),
+    (err: Error) =>
+      callback(denoErrorToNodeError(err, { syscall: "open", path })),
+  );
+}
+
+function openSync(path: string | Buffer | URL): number;
+function openSync(
+  path: string | Buffer | URL,
+  flags?: OpenFlags,
+): number;
+function openSync(path: string | Buffer | URL, mode?: number): number;
+function openSync(
+  path: string | Buffer | URL,
+  flags?: OpenFlags,
+  mode?: number,
+): number;
+function openSync(
+  path: string | Buffer | URL,
+  flags: OpenFlags = "r",
+  maybeMode?: number,
+) {
+  path = getValidatedPathToString(path);
+  flags = stringToFlags(flags);
+  const mode = parseFileMode(maybeMode, "mode", 0o666);
+
+  try {
+    return op_node_open_sync(path, flags, mode);
+  } catch (err) {
+    throw denoErrorToNodeError(err as Error, { syscall: "open", path });
+  }
+}
+
+// -- opendir --
+
+type OpendirOptions = {
+  encoding?: string;
+  bufferSize?: number;
+};
+type OpendirCallback = (err?: Error | null, dir?: Dir) => void;
+
+function _opendirValidateFunction(
+  callback: unknown,
+): asserts callback is OpendirCallback {
+  validateFunction(callback, "callback");
+}
+
+function _opendirGetPathString(
+  path: string | Buffer | URL,
+): string {
+  if (Buffer.isBuffer(path)) {
+    // deno-lint-ignore prefer-primordials
+    return path.toString();
+  }
+
+  return StringPrototypeToString(path);
+}
+
+function opendir(
+  path: string | Buffer | URL,
+  options: OpendirOptions | OpendirCallback,
+  callback?: OpendirCallback,
+) {
+  callback = typeof options === "function" ? options : callback;
+  _opendirValidateFunction(callback);
+
+  path = _opendirGetPathString(getValidatedPath(path));
+
+  let err, dir;
+  try {
+    const { bufferSize } = getOptions(options, {
+      encoding: "utf8",
+      bufferSize: 32,
+    });
+    validateInteger(bufferSize, "options.bufferSize", 1, 4294967295);
+
+    /** Throws if path is invalid */
+    Deno.readDirSync(path);
+
+    dir = new Dir(path);
+  } catch (error) {
+    err = denoErrorToNodeError(error as Error, { syscall: "opendir" });
+  }
+  if (err) {
+    callback(err);
+  } else {
+    callback(null, dir);
+  }
+}
+
+function opendirSync(
+  path: string | Buffer | URL,
+  options?: OpendirOptions,
+): Dir {
+  path = _opendirGetPathString(getValidatedPath(path));
+
+  const { bufferSize } = getOptions(options, {
+    encoding: "utf8",
+    bufferSize: 32,
+  });
+
+  validateInteger(bufferSize, "options.bufferSize", 1, 4294967295);
+
+  try {
+    /** Throws if path is invalid */
+    Deno.readDirSync(path);
+
+    return new Dir(path);
+  } catch (err) {
+    throw denoErrorToNodeError(err as Error, { syscall: "opendir" });
   }
 }
