+++ tools/core_import_map.json
@@ -540,9 +540,6 @@
     "ext:deno_node/_fs/_fs_ftruncate.ts": "../ext/node/polyfills/_fs/_fs_ftruncate.ts",
     "ext:deno_node/_fs/_fs_glob.ts": "../ext/node/polyfills/_fs/_fs_glob.ts",
     "ext:deno_node/_fs/_fs_lstat.ts": "../ext/node/polyfills/_fs/_fs_lstat.ts",
-    "ext:deno_node/_fs/_fs_mkdir.ts": "../ext/node/polyfills/_fs/_fs_mkdir.ts",
-    "ext:deno_node/_fs/_fs_open.ts": "../ext/node/polyfills/_fs/_fs_open.ts",
-    "ext:deno_node/_fs/_fs_opendir.ts": "../ext/node/polyfills/_fs/_fs_opendir.ts",
     "ext:deno_node/_fs/_fs_read.ts": "../ext/node/polyfills/_fs/_fs_read.ts",
     "ext:deno_node/_fs/_fs_readlink.ts": "../ext/node/polyfills/_fs/_fs_readlink.ts",
     "ext:deno_node/_fs/_fs_readv.ts": "../ext/node/polyfills/_fs/_fs_readv.ts",
